+++ src/rpc/quorums.cpp
@@ -164,7 +164,7 @@ static RPCHelpMan quorum_list_extended()
                 }
                 j.pushKV("creationHeight", q->m_quorum_base_block_index->nHeight);
                 j.pushKV("minedBlockHash", q->minedBlockHash.ToString());
-                j.pushKV("numValidMembers", (int32_t)num_valid_members);
+                j.pushKV("numValidMembers", num_valid_members);
                 j.pushKV("healthRatio", ss.str());
                 obj.pushKV(q->qc->quorumHash.ToString(),j);
             }
@@ -916,7 +916,7 @@ static RPCHelpMan quorum_dkginfo()
     llmq::CDKGDebugStatus status;
     llmq_ctx.dkg_debugman->GetLocalDebugStatus(status);
     UniValue ret(UniValue::VOBJ);
-    ret.pushKV("active_dkgs", int(status.sessions.size()));
+    ret.pushKV("active_dkgs", status.sessions.size());
 
     const int nTipHeight{WITH_LOCK(cs_main, return chainman.ActiveChain().Height())};
     auto minNextDKG = [](const Consensus::Params& consensusParams, int nTipHeight) {
